refactor(inbox): split PlanPanel.tsx into focused modules
Pure structural split, zero behaviour change. Extracts the per-group
row markup into PlanGroupRow, the destination-root prompt into
PlanRootPicker, the destructive-destination control into
PlanDestructiveControl, and the pure frame-type/summary helpers into
planPanelHelpers. PlanPanel.tsx keeps its public API (component,
props, and type exports) unchanged.
